refactor(#3280): extract `key in obj` operator from compileBinaryExpression (WAVE C slice 2)
Lift the self-contained `key in obj` relational-operator block (the `op ===
InKeyword` branch) out of compileBinaryExpression into a new sibling module
binary-ops-in.ts / compileInOperator. Handles: private-brand runtime check
(#1365), `in` on a statically-primitive RHS (§13.10.1 TypeError, #2741), vec
index bounds check, static/dynamic key resolution against struct fields + the
TS type system, and the `__extern_has` host MOP route (#1444).

- The block reads only ctx/fctx/expr (no operand-ValType coupling) and returns
  on every path, so the lift is a single guarded call. inRhsIsExclusivelyPrimitive
  promoted to export (no emit change).
- binary-ops.ts 3,164 -> 2,789 LOC; compileBinaryExpression ~1,710 -> ~1,330
  LOC (< 1,500 threshold — WAVE C goal met). New module 417 LOC.

Byte-identical: prove-emit-identity check IDENTICAL 39/39 (gc/standalone/wasi).
tsc 0, biome lint clean, smoke test tests/issue-3280.test.ts 11/11 green.

Stacked on slice 1 (PR #3090); completes #3280.


✓
