+++ src/codegen/binary-ops.ts
@@ -45,6 +45,7 @@ import { compileStringBinaryOp, emitHoistedCharCodeAtRead, matchHoistedCharRead 
 import { emitAnyEqFromExternTemps, emitLooseEq, emitStrictEq } from "./coercion-engine.js";
 import { compileInstanceOf, compileTypeofComparison } from "./typeof-delete.js";
 import { compileTypedBinaryDispatch } from "./binary-ops-typed-dispatch.js";
+import { compileInOperator } from "./binary-ops-in.js";
 
 // ── Binary operations ─────────────────────────────────────────────────
 
@@ -101,7 +102,7 @@ export function brandBooleanBinaryResult(op: ts.SyntaxKind, result: InnerResult)
  * object types and any union with a non-primitive constituent return false (they
  * defer to the runtime `[[HasProperty]]` / `__extern_has` check).
  */
-function inRhsIsExclusivelyPrimitive(t: ts.Type): boolean {
+export function inRhsIsExclusivelyPrimitive(t: ts.Type): boolean {
   const PRIM =
     ts.TypeFlags.Number |
     ts.TypeFlags.NumberLiteral |
@@ -592,387 +593,7 @@ export function compileBinaryExpression(
 
   // `key in obj` — compile-time property existence check
   if (op === ts.SyntaxKind.InKeyword) {
-    // #1365 — `#x in obj` is a RUNTIME brand check, not a compile-time
-    // property-name lookup. Per ES2022 §12.10.3 (RelationalExpression :
-    // PrivateIdentifier `in` ShiftExpression), the result is `true` iff
-    // `obj` carries the brand of the class that lexically declared `#x`,
-    // and `false` otherwise (no throw, even when obj isn't an object).
-    //
-    // Today the generic `in` path returns a compile-time `i32.const` based
-    // on whether the receiver type's struct happens to have `__priv_<name>`
-    // as a field. That conflates two unrelated classes both declaring a
-    // private named the same — `#x in instanceOfDifferentClass` returns
-    // true when it should return false.
-    //
-    // Fix: emit a runtime `ref.test` against the declaring class's struct.
-    // Falls through to the legacy path if the resolver can't find the
-    // declaring class (defensive — well-formed source always finds it).
-    if (ts.isPrivateIdentifier(expr.left)) {
-      const declared = resolveDeclaringClassForPrivateName(ctx, expr.left);
-      if (declared) {
-        // Compile the receiver. Coerce externref → anyref and save it so
-        // the brand predicate can combine structural ref.test with class-tag
-        // ancestry.
-        const objResult = compileExpression(ctx, fctx, expr.right);
-        if (objResult?.kind === "externref") {
-          fctx.body.push({ op: "any.convert_extern" });
-        }
-        const tmpAny = allocTempLocal(fctx, { kind: "anyref" });
-        fctx.body.push({ op: "local.set", index: tmpAny });
-        emitPrivateBrandPredicate(ctx, fctx, tmpAny, declared.className, declared.structTypeIdx);
-        releaseTempLocal(fctx, tmpAny);
-        return { kind: "i32" };
-      }
-      // No declaring class found — fall through to the legacy compile-time
-      // path. The compile-time bool will be wrong but at least won't trap.
-    }
-
-    const rightType = ctx.checker.getTypeAtLocation(expr.right);
-    let rightWasm = resolveWasmType(ctx, rightType);
-
-    // (#2741) §13.10.1 step 5 — `key in rval` throws a **TypeError** when
-    // `Type(rval)` is not Object. When the RHS static type is EXCLUSIVELY a
-    // non-object primitive (number / string / boolean / bigint / symbol / null /
-    // undefined, or a literal/union thereof), its runtime value can never be an
-    // Object, so emit a runtime throw rather than statically folding to a boolean
-    // (which is what the path below would do, e.g. `"toString" in true → true`).
-    // Spec evaluation order (steps 1-4): evaluate the LHS (key) then the RHS for
-    // side effects, THEN throw. `any` / `unknown` / object / `never` / a union
-    // containing a non-primitive constituent are NOT caught here — they defer to
-    // the runtime [[HasProperty]] / `__extern_has` path, which throws for a
-    // genuinely-primitive runtime value via the native `key in obj`.
-    if (inRhsIsExclusivelyPrimitive(rightType)) {
-      const lt = compileExpression(ctx, fctx, expr.left);
-      if (lt !== null) fctx.body.push({ op: "drop" });
-      const rt = compileExpression(ctx, fctx, expr.right);
-      if (rt !== null) fctx.body.push({ op: "drop" });
-      emitThrowTypeError(ctx, fctx, "Cannot use 'in' operator to search for property in a non-object");
-      return { kind: "i32" };
-    }
-
-    // (#2617) The TS type of a `new Proxy(...)`-bound identifier is its TARGET
-    // type (ProxyConstructor returns T), so `resolveWasmType` yields the target
-    // struct and the static `in` fold below would constant-fold `'k' in p` to
-    // the target's field membership — never calling `__extern_has`, so a `has`
-    // trap (incl. one that throws, #2617) never runs. But #2615 slots that
-    // variable as `externref`. Trust the ACTUAL slot type: if the receiver is an
-    // identifier whose local slot is externref/anyref, treat the RHS as externref
-    // so the `in` routes through `__extern_has` (the host Proxy MOP).
-    if (
-      (rightWasm.kind === "ref" || rightWasm.kind === "ref_null") &&
-      ts.isIdentifier(expr.right) &&
-      fctx.localMap.has(expr.right.text)
-    ) {
-      const idx = fctx.localMap.get(expr.right.text)!;
-      const entry = idx < fctx.params.length ? fctx.params[idx] : fctx.locals[idx - fctx.params.length];
-      const slotType =
-        entry && typeof entry === "object" && "type" in entry
-          ? (entry as { type: ValType }).type
-          : (entry as ValType | undefined);
-      if (slotType?.kind === "externref" || slotType?.kind === "anyref") {
-        rightWasm = slotType;
-      }
-    }
-
-    // Get struct field names if available; detect vec (array) types
-    let structFieldNames: string[] | null = null;
-    let isVecType = false;
-    let vecTypeIdx = -1;
-    if (rightWasm.kind === "ref" || rightWasm.kind === "ref_null") {
-      const typeIdx = (rightWasm as { typeIdx: number }).typeIdx;
-      const structDef = ctx.mod.types[typeIdx];
-      if (structDef?.kind === "struct") {
-        if (structDef.name?.startsWith("__vec_")) {
-          isVecType = true;
-          vecTypeIdx = typeIdx;
-        } else {
-          structFieldNames = structDef.fields.map((f) => f.name).filter((n): n is string => n !== undefined);
-        }
-      }
-    }
-
-    // Resolve the key to a compile-time string if possible.
-    // For comma expressions like (x = y, "key"), extract the last element.
-    // For PrivateIdentifier (#field in obj), extract the field name without '#'.
-    let staticKey: string | null = null;
-    const leftExpr: ts.Expression = expr.left;
-    if (ts.isPrivateIdentifier(leftExpr)) {
-      staticKey = leftExpr.text.startsWith("#") ? "__priv_" + leftExpr.text.slice(1) : leftExpr.text;
-    } else if (ts.isStringLiteral(leftExpr)) {
-      staticKey = leftExpr.text;
-    } else if (ts.isNumericLiteral(leftExpr)) {
-      staticKey = leftExpr.text;
-    } else if (ts.isBinaryExpression(leftExpr) && leftExpr.operatorToken.kind === ts.SyntaxKind.CommaToken) {
-      // Comma expression: extract the last element for the static key
-      let last: ts.Expression = leftExpr.right;
-      while (ts.isBinaryExpression(last) && last.operatorToken.kind === ts.SyntaxKind.CommaToken) {
-        last = last.right;
-      }
-      if (ts.isStringLiteral(last)) {
-        staticKey = last.text;
-      } else if (ts.isNumericLiteral(last)) {
-        staticKey = last.text;
-      }
-    } else if (ts.isParenthesizedExpression(leftExpr)) {
-      // Parenthesized expression: unwrap and check for comma or literal
-      const inner = leftExpr.expression;
-      if (ts.isBinaryExpression(inner) && inner.operatorToken.kind === ts.SyntaxKind.CommaToken) {
-        let last: ts.Expression = inner.right;
-        while (ts.isBinaryExpression(last) && last.operatorToken.kind === ts.SyntaxKind.CommaToken) {
-          last = last.right;
-        }
-        if (ts.isStringLiteral(last)) {
-          staticKey = last.text;
-        } else if (ts.isNumericLiteral(last)) {
-          staticKey = last.text;
-        }
-      } else if (ts.isStringLiteral(inner)) {
-        staticKey = inner.text;
-      } else if (ts.isNumericLiteral(inner)) {
-        staticKey = inner.text;
-      }
-    }
-
-    // Also check the TypeScript type system for property existence.
-    // This handles built-in constructors (Number.MAX_VALUE), prototype methods
-    // (valueOf, toString), and dynamically assigned properties.
-    let tsTypeHasProperty = false;
-    if (staticKey !== null) {
-      // Check direct properties on the TypeScript type
-      const prop = rightType.getProperty(staticKey);
-      if (prop) {
-        tsTypeHasProperty = true;
-      }
-      // Check the right side's type for comma expressions too
-      if (
-        !tsTypeHasProperty &&
-        ts.isBinaryExpression(expr.right) &&
-        expr.right.operatorToken.kind === ts.SyntaxKind.CommaToken
-      ) {
-        let lastRight: ts.Expression = expr.right.right;
-        while (ts.isBinaryExpression(lastRight) && lastRight.operatorToken.kind === ts.SyntaxKind.CommaToken) {
-          lastRight = lastRight.right;
-        }
-        const lastRightType = ctx.checker.getTypeAtLocation(lastRight);
-        const prop2 = lastRightType.getProperty(staticKey);
-        if (prop2) tsTypeHasProperty = true;
-      }
-      // Also check apparent type (includes prototype methods like valueOf, toString)
-      if (!tsTypeHasProperty) {
-        const apparentType = ctx.checker.getApparentType(rightType);
-        const apparentProp = apparentType.getProperty(staticKey);
-        if (apparentProp) tsTypeHasProperty = true;
-      }
-    }
-
-    // Array (vec) index bounds check: `index in arr` → 0 <= index < arr.length
-    if (isVecType && staticKey !== null) {
-      const numIdx = Number(staticKey);
-      if (Number.isFinite(numIdx) && numIdx >= 0 && Number.isInteger(numIdx)) {
-        // Evaluate left for side effects, drop result
-        const leftResult = compileExpression(ctx, fctx, expr.left);
-        if (leftResult) {
-          fctx.body.push({ op: "drop" });
-        }
-        // Compile the array expression to get the vec struct
-        const rightResult = compileExpression(ctx, fctx, expr.right);
-        if (rightResult) {
-          // Read length field (field 0 of vec struct)
-          fctx.body.push({ op: "struct.get", typeIdx: vecTypeIdx, fieldIdx: 0 });
-          // Compare: numIdx < length
-          fctx.body.push({ op: "i32.const", value: numIdx });
-          fctx.body.push({ op: "i32.gt_s" }); // length > index  <==>  index < length
-        } else {
-          fctx.body.push({ op: "i32.const", value: 0 });
-        }
-        return { kind: "i32" };
-      }
-      // Non-numeric key like "length" on array — check TS type
-      if (staticKey === "length") {
-        const leftResult = compileExpression(ctx, fctx, expr.left);
-        if (leftResult) {
-          fctx.body.push({ op: "drop" });
-        }
-        const rightResult = compileExpression(ctx, fctx, expr.right);
-        if (rightResult) {
-          fctx.body.push({ op: "drop" });
-        }
-        fctx.body.push({ op: "i32.const", value: 1 });
-        return { kind: "i32" };
-      }
-    }
-
-    // Static resolution: key is known at compile time
-    if (staticKey !== null) {
-      const hasInStruct = structFieldNames !== null && structFieldNames.includes(staticKey);
-      const has = hasInStruct || tsTypeHasProperty;
-      // (#1444) When RHS is externref/anyref AND static analysis came up empty
-      // (no struct field, no TS-typed prop), the answer is NOT reliably false
-      // — the host object may carry dynamic keys (e.g. regex `result.groups`).
-      // Route through `__extern_has` for the real `in` check instead of
-      // emitting an unconditional `false`.
-      if (!has && (rightWasm.kind === "externref" || rightWasm.kind === "anyref")) {
-        const hasIdx = ensureLateImport(
-          ctx,
-          "__extern_has",
-          [{ kind: "externref" }, { kind: "externref" }],
-          [{ kind: "i32" }],
-        );
-        if (hasIdx !== undefined) {
-          flushLateImportShifts(ctx, fctx);
-          // (#2741) §13.10.1 evaluates the LHS (key, steps 1-2) BEFORE the RHS
-          // (object, steps 3-4). Evaluate the key first into a temp, then the
-          // object, then re-push the key so the call args are `(obj, key)`.
-          // Use coerceType (not a bare extern.convert_any) so a non-ref key
-          // (e.g. `Infinity` → f64) is boxed to externref via __box_number.
-          const leftResult = compileExpression(ctx, fctx, expr.left, { kind: "externref" });
-          if (leftResult === null) {
-            fctx.body.push({ op: "ref.null.extern" });
-          } else if (leftResult.kind !== "externref") {
-            coerceType(ctx, fctx, leftResult, { kind: "externref" });
-          }
-          const keyTmp = allocTempLocal(fctx, { kind: "externref" });
-          fctx.body.push({ op: "local.set", index: keyTmp });
-          const rightResult = compileExpression(ctx, fctx, expr.right, { kind: "externref" });
-          if (rightResult === null) {
-            fctx.body.push({ op: "ref.null.extern" });
-          } else if (rightResult.kind !== "externref") {
-            coerceType(ctx, fctx, rightResult, { kind: "externref" });
-          }
-          fctx.body.push({ op: "local.get", index: keyTmp });
-          releaseTempLocal(fctx, keyTmp);
-          fctx.body.push({ op: "call", funcIdx: hasIdx });
-          return { kind: "i32" };
-        }
-      }
-      // Evaluate both operands for side effects (needed for comma expressions like
-      // (NUMBER = Number, "MAX_VALUE") in NUMBER). Drop the produced values.
-      const leftResult = compileExpression(ctx, fctx, expr.left);
-      if (leftResult) {
-        fctx.body.push({ op: "drop" });
-      }
-      const rightResult = compileExpression(ctx, fctx, expr.right);
-      if (rightResult) {
-        fctx.body.push({ op: "drop" });
-      }
-      fctx.body.push({ op: "i32.const", value: has ? 1 : 0 });
-      return { kind: "i32" };
-    }
-
-    // Dynamic key with known struct fields: runtime string comparison.
-    // (#2741) Gate to a REFERENCE-like key (string / externref / anyref). A
-    // value-typed key (`Infinity`/`true`/a number → f64/i32) cannot be fed to
-    // `__str_eq` (it expects a string/externref) — doing so produced a malformed
-    // module ("call expected externref, found f64"). Such keys (now reachable
-    // because the §13.10.1 ToPropertyKey 2322 is downgraded) fall through to the
-    // defined fallback below instead of crashing wasm validation.
-    const leftKeyWasm = resolveWasmType(ctx, ctx.checker.getTypeAtLocation(expr.left));
-    const keyIsRefLike =
-      leftKeyWasm.kind === "externref" ||
-      leftKeyWasm.kind === "anyref" ||
-      leftKeyWasm.kind === "ref" ||
-      leftKeyWasm.kind === "ref_null";
-    if (structFieldNames !== null && structFieldNames.length > 0 && keyIsRefLike) {
-      // Compile the key expression (should produce a string/externref)
-      const keyType = compileExpression(ctx, fctx, expr.left);
-      if (keyType) {
-        // Compare key against each field name using wasm:js-string equals
-        const equalsIdx = ctx.funcMap.get("__str_eq") ?? ctx.funcMap.get("string_equals");
-        const jsStrEquals = ctx.mod.imports.findIndex(
-          (imp) => imp.module === "wasm:js-string" && imp.name === "equals",
-        );
-        const eqFunc = jsStrEquals >= 0 ? jsStrEquals : equalsIdx;
-        if (eqFunc !== undefined && eqFunc >= 0) {
-          const keyLocal = allocLocal(fctx, `__in_key_${fctx.locals.length}`, keyType);
-          fctx.body.push({ op: "local.set", index: keyLocal });
-          // Start with false (0)
-          fctx.body.push({ op: "i32.const", value: 0 });
-          for (const fieldName of structFieldNames) {
-            // Load the key and the field name string, compare
-            fctx.body.push({ op: "local.get", index: keyLocal });
-            const strGlobal = ctx.stringGlobalMap.get(fieldName);
-            if (strGlobal !== undefined) {
-              fctx.body.push({ op: "global.get", index: strGlobal });
-              fctx.body.push({ op: "call", funcIdx: eqFunc });
-              fctx.body.push({ op: "i32.or" }); // OR with accumulated result
-            }
-          }
-          return { kind: "i32" };
-        }
-      }
-    }
-
-    // Dynamic key with no struct fields — try TS type system for known properties
-    // Compile both sides for side effects, then use TS type system if the key
-    // can be resolved from its type (e.g., a string variable with a known literal type).
-    {
-      // (#1444) When RHS is externref-backed (host object — e.g. regex
-      // `result.groups`, untyped JS values), route through `__extern_has` so
-      // `'key' in hostObj` reflects the actual JS `in` semantics instead of
-      // the unconditional `false` fallback. The static path above still
-      // covers WasmGC structs / vec types / TS-typed properties where the
-      // compile-time answer is reliable.
-      if (rightWasm.kind === "externref" || rightWasm.kind === "anyref") {
-        const hasIdx = ensureLateImport(
-          ctx,
-          "__extern_has",
-          [{ kind: "externref" }, { kind: "externref" }],
-          [{ kind: "i32" }],
-        );
-        if (hasIdx !== undefined) {
-          flushLateImportShifts(ctx, fctx);
-          // (#2741) §13.10.1 evaluates the LHS (key, steps 1-2) BEFORE the RHS
-          // (object, steps 3-4) — e.g. `x() in y()` must throw from `x()` first,
-          // and an unresolvable LHS reference (`undef in obj`) must throw before
-          // the object is evaluated. Evaluate the key first into a temp, then the
-          // object, then re-push the key so the call args stay `(obj, key)`.
-          // coerceType (not a bare extern.convert_any) boxes a non-ref key.
-          const leftResult = compileExpression(ctx, fctx, expr.left, { kind: "externref" });
-          if (leftResult === null) {
-            fctx.body.push({ op: "ref.null.extern" });
-          } else if (leftResult.kind !== "externref") {
-            coerceType(ctx, fctx, leftResult, { kind: "externref" });
-          }
-          const keyTmp = allocTempLocal(fctx, { kind: "externref" });
-          fctx.body.push({ op: "local.set", index: keyTmp });
-          const rightResult = compileExpression(ctx, fctx, expr.right, { kind: "externref" });
-          if (rightResult === null) {
-            fctx.body.push({ op: "ref.null.extern" });
-          } else if (rightResult.kind !== "externref") {
-            coerceType(ctx, fctx, rightResult, { kind: "externref" });
-          }
-          fctx.body.push({ op: "local.get", index: keyTmp });
-          releaseTempLocal(fctx, keyTmp);
-          fctx.body.push({ op: "call", funcIdx: hasIdx });
-          return { kind: "i32" };
-        }
-      }
-
-      const leftResult = compileExpression(ctx, fctx, expr.left);
-      if (leftResult) {
-        fctx.body.push({ op: "drop" });
-      }
-      const rightResult = compileExpression(ctx, fctx, expr.right);
-      if (rightResult) {
-        fctx.body.push({ op: "drop" });
-      }
-
-      // Try to resolve key from the TS type of the left expression
-      const leftType = ctx.checker.getTypeAtLocation(expr.left);
-      if (leftType.isStringLiteral()) {
-        const key = leftType.value;
-        const prop = rightType.getProperty(key);
-        const apparentType = ctx.checker.getApparentType(rightType);
-        const apparentProp = apparentType.getProperty(key);
-        const has = !!(prop || apparentProp || (structFieldNames && structFieldNames.includes(key)));
-        fctx.body.push({ op: "i32.const", value: has ? 1 : 0 });
-        return { kind: "i32" };
-      }
-
-      // Fully dynamic — emit false as safe fallback
-      fctx.body.push({ op: "i32.const", value: 0 });
-      return { kind: "i32" };
-    }
+    return compileInOperator(ctx, fctx, expr);
   }
 
   // ── Flatten long chains of same numeric operator ──
